feat: Introduce ModeToggle component and integrate it into CollectionTreesView, enhancing UI with improved display modes
