+++ view/desktop/src/parts/HeadBar/CollapsibleActionMenu.tsx
@@ -6,8 +6,6 @@ import PanelToggleButtons from "./PanelToggleButtons";
 
 export interface CollapsibleActionMenuProps {
   isCompact?: boolean;
-  showDebugPanels: boolean;
-  setShowDebugPanels: (show: boolean) => void;
   openPanel: (panel: string) => void;
 }
 
@@ -15,17 +13,10 @@ export interface CollapsibleActionMenuProps {
 export const CollapsibleActionMenu = ({ isCompact = false }: CollapsibleActionMenuProps) => {
   const { sideBarPosition, bottomPane, sideBar } = useAppResizableLayoutStore();
 
-  // When not in compact mode, show all buttons
   if (!isCompact) {
-    return (
-      <div className="flex items-center gap-0">
-        <PanelToggleButtons className="mr-1" />
-        <ActionButton icon="Bell" iconClassName="text-(--moss-headBar-icon-primary-text)" />
-      </div>
-    );
+    return <PanelToggleButtons />;
   }
 
-  // In compact mode, use ActionMenu
   return (
     <ActionMenu.Root>
       <ActionMenu.Trigger asChild>
@@ -36,9 +27,6 @@ export const CollapsibleActionMenu = ({ isCompact = false }: CollapsibleActionMe
         />
       </ActionMenu.Trigger>
       <ActionMenu.Content>
-        <ActionMenu.Item onClick={() => {}} icon="Bell">
-          Notifications
-        </ActionMenu.Item>
         {sideBarPosition === SIDEBAR_POSITION.LEFT ? (
           <>
             <ActionMenu.Item
