+++ view/desktop/src/parts/StatusBar/StatusBarActivity.tsx
@@ -168,7 +168,7 @@ export const StatusBarActivity = () => {
 
   return (
     <div className="flex h-full items-center">
-      <button className="group flex h-full items-center rounded px-2 transition">
+      <button className="group flex h-full items-center rounded transition">
         <div className="hover:background-(--moss-statusBar-icon-background-hover) flex h-[22px] items-center gap-1.5 rounded px-1">
           <Icon
             className={cn(
