+++ view/desktop/src/parts/StatusBar/ZoomButtons.tsx
